Part 1-9: From Setup to Updating User Profile
Part of the Corey Schafer Youtube series on Django. This initial commit already contains all the files for the first 9 videos. At this stage the site contains the ability to login, logout, create accounts, view a profile, etc. Three models are used: One for users (already in django by default), one for profiles (additional user info) and one for posts.