Part 10 - Create, Update and Delete Posts
Migrated to the use of class based views, and utilized the default classes for listing, deleting, updating and creating, to implement functionality around the manipulation of posts. Also added the ability to only allow the creation of posts for logged in users, and only allow to update a post if the current logged in user is the author.