Part 12 - Email and Password Reset
Added the ability for users to change their passwords. Gmail was used to do this, and an app specific password allowed the access of gmail through django. 4 different routes are needed, but everything else is handled by django. Just be careful with the names of the routes. If they dont have underscores, the underlying class used to reset the passwords will have trouble finding the appropriate routes.