Part 13 and 14 - Deployment and S3
To deploy the website onto Heroku, we have to make the django app use S3 buckets. After that, in part 14, many changes were implemented in order for the deployment to properly work, while retaining the ability to modify the website locally. The entire process was documented inside the Note.md file.